modify: 방 만들기 모달로 변경
