+++ client/src/components/modals/createRoom/createRoom.module.scss
@@ -1,20 +1,7 @@
-@import "../../styles/variables";
-@import "../../styles/mixin";
+@import "../../../styles/mixin";
 
 .modal_container {
-    display: block;
-    position: absolute;
-    width: 350px;
-    height: 240px;
-    padding: 30px 0 30px;
-    border-radius: 15px;
-    border: 1px solid $gray-light;
-    text-align: center;
-    background: $white;
-    box-shadow: 0px 0px 24px -4px rgba(16, 24, 40, 0.08);
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    @include modal_container;
 }
 
 .close_icon {
@@ -36,6 +23,7 @@
     width: 250px;
     height: 38px;
     margin: 30px 0 37px;
+    padding: 8px 12px 8px 16px;
     border: 1px solid $point-color;
     text-align: left;
 }
